llm-token-monitor/requirements.txt
fastapi==0.118.3
uvicorn==0.37.0
openai==1.108.1
prometheus_client==0.23.1
pydantic==2.11.9
python-dotenv==1.1.1